+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 Popmotion adheres to [Semantic Versioning](http://semver.org/).
 
+## [6.3.5] 2017-03-03
+
+## Added
+- `applyOffset` transformer.
+
+## [6.3.4] 2017-03-01
+
+## Fixed
+- Undocumented range easing property in `interpolate` transformer was looking at the wrong index.
+
+## [6.3.3] 2017-18-01
+
+### Added
+- `top`, `left`, `bottom` and `right` CSS renderer value types.
+
+## [6.3.2] 2017-11-01
+
+### Added
+- `smooth` transformer
+
+## [6.3.1] 2017-11-01
+
+### Fixed
+- `transformChildValues` transformer now keeps a local mutable state rather than mutating the provided state.
+
+## [6.3.0] 2017-11-01
+
+### Added
+- Actions now have an `output` shorthand for setting the `onUpdate` property.
+
 ## [6.2.1] 2017-09-01
 
 ### Added